backend/curso_administracao/requirements.txt
asgiref
astroid
autopep8
cffi
cryptography
certifi
chardet
coreapi
coreschema
dj-static==0.0.6
Django==3.2.12
django-absoluteuri
django-adminlte3
django-cors-headers
django-rest-swagger==2.2.0
django-sql-utils
django-stdimage
djangorestframework
djangorestframework-jwt
djangorestframework-simplejwt
Markdown
django-filter
drf-yasg==1.20.0
gunicorn==20.0.4
idna==2.10
importlib-metadata==1.7.0
inflection==0.5.1
isort==4.3.21
jwcrypto==0.8
itypes==1.2.0
Jinja2==2.11.3
lazy-object-proxy==1.4.3
MarkupSafe==1.1.1
mccabe==0.6.1
numpy
openapi-codec==1.3.2
packaging==20.9
pandas
Pillow==7.2.0
progressbar2==3.51.4
psycopg2-binary==2.8.5
pycodestyle==2.6.0
pycparser==2.20
PyJWT==1.7.1
pylint==2.5.3
pylint-django==2.0.15
pylint-plugin-utils==0.6
pyparsing==2.4.7
python-dateutil==2.8.1
python-utils==2.4.0
pytz==2020.1
requests==2.25.1
ruamel.yaml==0.16.13
ruamel.yaml.clib==0.2.2
simplejson==3.17.2
six==1.15.0
sqlparse==0.3.1
static3==0.7.0
toml==0.10.1
typed-ast==1.4.1
uritemplate==3.0.1
urllib3==1.26.4
wrapt==1.12.1
xlrd==1.2.0
zipp==3.1.0


